requirements/test.txt
coverage==6.3.2
diff_cover==5.1.2
django-sslserver==0.20
freezegun==0.3.12
github3.py==0.9.6
model-bakery==1.1.0
moto==1.3.6
pip==21.1.2
responses==0.9.0
wagtail-foliage==0.1.1
